removed code smells and duplications
