+++ src/models/TracklineDonor/tracklineDonor.ts
@@ -1,5 +1,4 @@
 import { FORM_ACTIONS } from "../../components/Forms/constant";
-import { DocumentNode } from "graphql";
 
 export type TracklineDonorFormProps = {
 	donors:
